+++ src/command/moveToBottom.ts
@@ -3,11 +3,12 @@ import * as vscode from 'vscode'
 import { Resource, FavoritesProvider } from '../provider/FavoritesProvider'
 import configMgr from '../helper/configMgr'
 import { getCurrentResources } from '../helper/util'
+import { DEFAULT_GROUP } from '../enum'
 
 export function moveToBottom(favoritesProvider: FavoritesProvider) {
   return vscode.commands.registerCommand('favorites.moveToBottom', async function (value: Resource) {
     const config = vscode.workspace.getConfiguration('favorites')
-    const currentGroup = configMgr.get('currentGroup') as string
+    const currentGroup = (configMgr.get('currentGroup') as string) || DEFAULT_GROUP
 
     const items = await getCurrentResources()
     const filteredArray: {